refactor(zebra-chain): update changelog and move tests
