+++ zebra-chain/CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+
+- Refactor `NetworkUpgrade` to use `strum::EnumIter` instead of `NETWORK_UPGRADES_IN_ORDER` and add the `strum` dependency to `zebra-chain`. (#10096)
+
 ## [4.0.0] - 2026-01-21
 
 ### Breaking Changes
