+++ zebra-chain/src/parameters/tests.rs
@@ -257,3 +257,36 @@ proptest! {
         }
     }
 }
+
+/// A list of network upgrades in the order that they must be activated.
+const NETWORK_UPGRADES_IN_ORDER: &[NetworkUpgrade] = &[
+    Genesis,
+    BeforeOverwinter,
+    Overwinter,
+    Sapling,
+    Blossom,
+    Heartwood,
+    Canopy,
+    Nu5,
+    Nu6,
+    Nu6_1,
+    #[cfg(any(test, feature = "zebra-test"))]
+    Nu7,
+];
+
+#[test]
+fn network_upgrade_iter_matches_order_constant() {
+    let iter_upgrades: Vec<NetworkUpgrade> = NetworkUpgrade::iter().collect();
+    let expected_upgrades: Vec<NetworkUpgrade> = NETWORK_UPGRADES_IN_ORDER.to_vec();
+
+    assert_eq!(iter_upgrades, expected_upgrades);
+}
+
+#[test]
+fn full_activation_list_contains_all_upgrades() {
+    let network = Network::Mainnet;
+    let full_list = network.full_activation_list();
+
+    // NU7 is only included in tests; on Mainnet, NU7 isn’t live yet, so we subtract 1 here.
+    assert_eq!(full_list.len(), NetworkUpgrade::iter().count() - 1);
+}
